pyproject.toml
[tool.poetry]
name = "api-workshop"
version = "0.1.0"
description = ""
authors = ["Luiz Fernando <luizfsoliveira.lm@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11 || ^3.12"
fastapi = "^0.115.6"
taskipy = "^1.14.1"
uvicorn = "^0.34.0"
pytest = "^8.3.4"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"



[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "uvicorn app.main:app --reload"
test = "pytest -v"
kill = "for /f \"tokens=5\" %a in ('netstat -aon ^| findstr :8000 ^| findstr LISTENING') do taskkill /F /PID %a"